Final typescript push.
Tried a number of different models and approaches to implement the tool, using both the transformers.js library and the hugging face API, but neither could deliver good results. Most of the models available for free in the API or in the transformers.js library are too small to produce an entire document.