Add extraction function for types and interfaces and add chalk stylings for console output
Added an extraction function for interfaces and types and amended generateDocs to account for this input.

Tool now generates docs for types and interfaces as well as functions. Amended the prompt to account for this.

Installed Chalk package and added stylings to console output.

Added more source code files to the files directory to input into the tool.