Re-add chunking and decrease limit to 2.
Re-added chunking for source code blocks that are sent in request to the LLM. Decreased the limit to 2 as when the source code blocks were too long, the model tended to skip over and omit large parts of the extracted symbols, leading to some functions not being documented in the output file. Decreasing the limit and the amount of code submitted per request seems to have solved this issue.

Also added statement to display what symbols are being generated for each chunk. Re-added the name property to extraction function output objects to do this.